Restore command uses gp_create_restore_point in QE
Updated the dispatch command to use gp_create_restore_point against
semgents. In the context of executors we will return segment_id and
restore_lsn using SRF as well. In the context of QD we will iterate
through the context tuples and add them to the QDs restore point tuple